Move Chaincode command into its own file
